Code duplicates with Glide optimized
