+++ app/src/main/java/space/taran/arknavigator/mvp/view/FoldersView.kt
@@ -3,7 +3,6 @@ package space.taran.arknavigator.mvp.view
 import moxy.MvpView
 import moxy.viewstate.strategy.AddToEndSingleStrategy
 import moxy.viewstate.strategy.StateStrategyType
-import space.taran.arknavigator.mvp.model.repo.Folders
 import java.nio.file.Path
 
 @StateStrategyType(AddToEndSingleStrategy::class)
